[Linux_PP] Changing default memory node setting
If the scheduling policy does not assign any memory node. By default
the Linux Platform Proxy will assign ALL the memory nodes. This
policy replaces the previous one, according to which the node set was
"0".

In a NUMA system the new policy is the behaviour to prefer also because
the old one can lead to performance issue in the application execution.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>