[YM] Fix wait for reconfigure before reclaiming resources
After an application receives a new resource allocation, it could choose to
terminate before doing the reconfiguration.
When the application terminates, it notifies it to the resource manager.

This is a deadlock: the resource manager is waiting for the app to notify
it a configuration happened, while the application is waiting for
the resource manager to ACK the application termination.

Removing the PostChange response solves this problem.

Signed-off-by: Simone Libutti <simone.libutti@gmail.com>