[Utils] Add a custom assert function
Add a custom assert function in order to be used in replace of std::assert.
The structure of Barbeque makes hard to immediately detect and debug a
false assert, since Barbeque is usually run in background.

The custom assert tries to write in the log, if not possible it writes
in stdout. After that an abort is called.

Signed-off-by: Federico Reghenzani <federico.reghenzani@polimi.it>