[RA] Fix error condition on remaining amount to assign
In case of BALANCED resource assignment policy, if happened that if was not
always possible to evenly distribute the assigned amount among N resources.
Therefore a small quantity remained enabling the condition 'requested!=0'
for which an error was returned. This misbehavior, experienced with
TEMPURA and 6 or 8 applications running on a 3-core configured platform.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>