[libpms] Resource allocation notification
On the programming model side it is required to wait for the resource
manager to return a resource allocation for the given application task
graph. To this purpose this commit introduces the
WaitForResourceAllocation() blocking function, which returns whenever the
NotifyResourceAllocation() is called. This happens in the onConfigure()
implementation, when the resource manager notifies that the allocation
has been performed.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>