[libpms] ExecSync: Task profiling implementation
First version of the task monitoring and profiling. For each running task
a task profiler is launched to collect timing information. The event
used for synchronization purpose is the one associated to the first
output buffer of the task. This allows us the get an estimation of the
task throughput.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>