[SM] Synchronization on the policy execution
Added the support to allows other modules to wait for the completion of
the current scheduling policy execution.

A state variable has been introduced to set keep track of the status of
the Scheduler Manager. When the state is set to SCHEDULING, a policy
execution is in progress, otherwise the state is READY.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>