[TRIG] Introducing trigger objects
A trigger is a component including boolean functions aimed at verify if,
given some input parameters, a certain condition is verified or not.
A typical use case is the monitoring of hardware resources status and the
detection of condition for which an execution of the optimization policy
must be triggered.

A first type of trigger introduced with this commit is one based on
threshold values.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>