[Daemon] Change PIL directory permissions
Enable write permissions to "all" users. This in order to allows the
installation of the predefined platform descriptions.
In fact, when the resource manager is launched in the daemon mode, the
starting script creates the PIL directory setting the ownership to 'root'.
Next attempts of installing platform description files would end in a
failure if suitable permissions are not set.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>