[DataManager] Definition of the DataManager class and inner classes
*The status data communication for monitoring between Barbeque and
thirdy-party applications is based on a publish-subscribe mechanism.
Thirdy-party applications can specify a specific status informations
(applications or resources status), and subscribe for a rate-based updating
(the status is published every X milliseconds) or an event-based updating
(the status is propagated on a specific event).
*The DataManager module is in charge of handling incoming subscriptions and
publish status updates to all the subscribers.
* The Subscription class models the filter, event and rate information of
each Subscriber.
* The Subscriber class models the single subscriber.

Signed-off-by: Michele Zanella <michele.zanella@polimi.it>