[DataManager] Implementation of the Unsubscribe method
The function takes a subscriber with a flag to indicate if it is
event-based.
Then it verifies if there is already a subscription, in this case it
updates the value of the filter, event or rate of the previous one.
If there are no more filters or events in the subscription it
removes the client from the proper subscribers list.

Signed-off-by: Michele Zanella <michele.zanella@polimi.it>