[libdci] First implementation of the DataClient library class
This library provide a class that is the base class for
third-party client to interface with the BarbequeRTRM DataManager module.
It hides from the developers the connection and network management.
It provides a method to subscribe to specific event and information filter.
It allows the client to set a callback function to update
the client when new status information are published by
BarbequeRTRM instance.
This class should be instantiated by the third-party client.

Signed-off-by: Michele Zanella <michele.zanella@polimi.it>