[libdci] Connection management refactoring
Connection is now performed in DataClient constructor, while disconnection
happens in destructor.

The constructor wait for the ClientReceiver thread to start and
successfully connect to the BarbequeRTRM server. A certain number of
attempts is provided.

In ClientReceiver the function iterates, until the connection is up,
by managing the incoming messages and executing the callback function.
Connection has been moved out of the loop.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>