[TRIG] Trigger refactoring
Deep refactoring of the trigger model.
* Improved execution mechanism of trigger by adding an arming threshold
(threshold_low) and an action threshold (threshold_high).
In this way to execute a trigger has to be armed to avoid that small
obscillations could cause a repeated action of the trigger.
* The Trigger object now contains information on its thresholds, margin
and armed status. A function IsArmed is provided to verify the status of
the armed flag. A constructor is provided to set all those parameters.
* The Trigger has a Check function that can be set customly and that
is used to verify if a condition is verified. Moreover it provided a
default Check function if no one is provided.
* The Trigger has an Action function that is set customly and that
is used to perform an action when the condition of the Check function is
verified and the trigger is activated.
* We introduced two type of threshold triggers, the OverThresholdTrigger
which is used to check value that exceeds a given threshold, and
the UnderThresholdTrigger which is used conversely.
* The TriggerFactory is in charge to provide one of the trigger types
given an id string argument.
* The bbque.config.in file is updated accorindgly to match config field
names changes.
* Code refactoring in the PowerMonitor code to match the changes in the
trigger model and functionalities.

Signed-off-by: Michele Zanella <michele.zanella@polimi.it>