[PRM] ProcessManager early version
This component as been thought to operate similarly to ApplicationManager,
but with Process objects. This in order to manage the allocation of
resources to generic not-integrated applications. In doing this, the
class provides member functions to add or remove processes by 'binary
name' and other ones to notify processes start and stop, according to
which the PRM should consider the triggering of the resource allocation
policy execution.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>