[RB] Fix binding when interleaved by added requests
This commit introduces two fixes:

1) When a resource binding has been already performed, and the policy goes on
by adding further resource requests, these last were ignored since they not
figured in the map of previous (partial) bindings.

2) The refinement of a resource path, through multiple bind calls was not
properly performed. The result was a continuous addition of bound assignments
to the map. With this commit the Bind() function removes the previous bind
assignments if they are compatible with the current one:
	e.g. "sys2.grp2.acc.pe" is removed if a binding to "sys2.grp2.acc3.pe"
        has been performed in the current function call

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>